Arregle el handshake del CPU a memoria
Cambie la funcion de void *buffer_leer_stream_recv por t_buffer *buffer_leer_recv que devulve el buffer, la primera funcion no existe mas.

+++ utils/src/utils/structs/structSendRecvMSG.h
@@ -41,7 +41,7 @@ typedef struct
 {
     TipoModulo moduloRemitente;
     TipoModulo moduloResponde;
-    uint8_t respuesta_cod;
+    uint32_t respuesta_cod;
 } t_resultHandShake;
 
 #endif
